Create Brute Force Calculator Generator.py
This simple python program generates C++ source code for a "brute force calculator." There is the old programming joke of a student who was tasked with making a calculator and spent all night making conditional statements for every possible combination of numbers he could manage. This program generates a rudimentary calculator that works just like that. It is not intended to be good; it was written as a joke.